- add save-source-in-map parameter - add target parameter

+++ source/class/qx/tool/cli/commands/Deploy.js
@@ -49,12 +49,25 @@ qx.Class.define("qx.tool.cli.commands.Deploy", {
             default: null,
             alias: "m"
           },
+          "save-source-in-map": {
+            describe: "Saves the source code in the map file (build target only)",
+            type: "boolean",
+            default: false
+          },
+    
           "clean": {
             describe: "Deletes the application output directory before deploying",
             type: "boolean",
             default: false,
             alias: "D"
-          }
+          },
+          "target": {
+            alias: "t",
+            describe: "Set the target type. Default is build",
+            type: "string",
+            default: "build"
+          },
+    
         }
       };
     }
@@ -68,7 +81,7 @@ qx.Class.define("qx.tool.cli.commands.Deploy", {
     process: async function() {
       let argv = this.argv;
       
-      let config = this.__config = await qx.tool.cli.Cli.getInstance().getParsedArgs();
+      let config  = await qx.tool.cli.Cli.getInstance().getParsedArgs();
       if (!config) {
         throw new qx.tool.utils.Utils.UserError("Error: Cannot find any configuration");
       }
@@ -81,7 +94,6 @@ qx.Class.define("qx.tool.cli.commands.Deploy", {
         qx.tool.compiler.Console.print("qx.tool.cli.deploy.notClean");
       }
       
-      config.targetType = "build";
       let compileArgv = {
         target: config.targetType,
         writeLibraryInfo: false,
@@ -90,6 +102,7 @@ qx.Class.define("qx.tool.cli.commands.Deploy", {
         saveUnminified: false,
         bundling: true,
         minify: "mangle",
+        saveSourceInMap: config.saveSourceInMap,
         __deploying: true
       };
       let appNames = null;
@@ -150,11 +163,13 @@ qx.Class.define("qx.tool.cli.commands.Deploy", {
         });
         {
           let from = path.join(target.getOutputDir(), "resource");
-          let to = path.join(argv.out, "resource");
-          if (makerIndex == 0 && argv.clean) {
-            await qx.tool.utils.files.Utils.deleteRecursive(to);
+          if (fs.existsSync(from)) {
+            let to = path.join(argv.out, "resource");
+            if (makerIndex == 0 && argv.clean) {
+              await qx.tool.utils.files.Utils.deleteRecursive(to);
+            }
+            await qx.tool.utils.files.Utils.sync(from, to);
           }
-          await qx.tool.utils.files.Utils.sync(from, to);
         }
         {
           let from = path.join(target.getOutputDir(), "index.html");
